+++ src/org/lukasj/idea/torquescript/completion/TSCompletionContributor.kt
@@ -7,12 +7,14 @@ import com.intellij.patterns.PlatformPatterns.psiElement
 import com.intellij.patterns.StandardPatterns
 import com.intellij.psi.PsiElement
 import com.intellij.psi.TokenType
+import com.intellij.psi.tree.TokenSet
 import org.lukasj.idea.torquescript.TSLanguage
 import org.lukasj.idea.torquescript.psi.TSTypes
 
 
 class TSCompletionContributor : CompletionContributor() {
     init {
+        extend(CompletionType.BASIC, inMethod(), TSMethodCompletionContributor())
         extend(CompletionType.BASIC, inDatablock(), TSClassCompletionContributor())
         extend(CompletionType.BASIC, inSingleton(), TSClassCompletionContributor())
         extend(CompletionType.BASIC, inNew(), TSClassCompletionContributor())
@@ -27,8 +29,8 @@ class TSCompletionContributor : CompletionContributor() {
 
     private fun isKeywordable(): ElementPattern<PsiElement> =
         psiElement(TSTypes.IDENT).withLanguage(TSLanguage.INSTANCE)
-            .andNot(psiElement().afterSibling(psiElement(TSTypes.DOT)))
-            .andNot(psiElement().afterSibling(psiElement(TSTypes.COLON_COLON)))
+            .andNot(psiElement().afterLeaf(psiElement(TSTypes.DOT)))
+            .andNot(psiElement().afterLeaf(psiElement(TSTypes.COLON_COLON)))
 
     private fun inGlobalCall(): ElementPattern<PsiElement> =
         isKeywordable()
@@ -36,53 +38,79 @@ class TSCompletionContributor : CompletionContributor() {
     private fun inGlobalNSCall(): ElementPattern<PsiElement> =
         psiElement(TSTypes.IDENT)
             .withLanguage(TSLanguage.INSTANCE)
-            .afterSibling(psiElement(TSTypes.COLON_COLON))
+            .afterLeaf(psiElement(TSTypes.COLON_COLON))
+            .andNot(
+                psiElement()
+                    .afterLeafSkipping(
+                        psiElement()
+                            .withElementType(
+                                TokenSet.create(
+                                    TokenType.WHITE_SPACE,
+                                    TSTypes.IDENT,
+                                    TSTypes.COLON_COLON
+                                )
+                            ),
+                        psiElement(TSTypes.FUNCTION)
+                    )
+            )
 
     private fun inMethodCall(): ElementPattern<PsiElement> =
         psiElement(TSTypes.IDENT)
             .withLanguage(TSLanguage.INSTANCE)
-            .afterSibling(psiElement(TSTypes.DOT))
+            .afterLeaf(psiElement(TSTypes.DOT))
 
     private fun inLocalVariable(): ElementPattern<PsiElement> =
         StandardPatterns.or(
             psiElement(TSTypes.IDENT)
-                .andNot(psiElement().afterSibling(psiElement(TSTypes.DOT)))
-                .andNot(psiElement().afterSibling(psiElement(TSTypes.COLON_COLON))),
+                .andNot(psiElement().afterLeaf(psiElement(TSTypes.DOT)))
+                .andNot(psiElement().afterLeaf(psiElement(TSTypes.COLON_COLON))),
             psiElement(TSTypes.LOCALVAR)
         )
 
     private fun inGlobalVariable(): ElementPattern<PsiElement> =
         StandardPatterns.or(
             psiElement(TSTypes.IDENT)
-                .andNot(psiElement().afterSibling(psiElement(TSTypes.DOT)))
-                .andNot(psiElement().afterSibling(psiElement(TSTypes.COLON_COLON))),
+                .andNot(psiElement().afterLeaf(psiElement(TSTypes.DOT)))
+                .andNot(psiElement().afterLeaf(psiElement(TSTypes.COLON_COLON))),
             psiElement(TSTypes.GLOBALVAR)
         )
 
     private fun inObjectName(): ElementPattern<PsiElement> =
         isKeywordable()
 
+    private fun inMethod(): ElementPattern<PsiElement> =
+        psiElement(TSTypes.IDENT)
+            .afterLeaf(psiElement(TSTypes.COLON_COLON))
+            .afterLeafSkipping(
+                psiElement()
+                    .withElementType(TokenSet.create(TokenType.WHITE_SPACE, TSTypes.IDENT, TSTypes.COLON_COLON)),
+                psiElement(TSTypes.FUNCTION)
+            )
+
     private fun inDatablock(): ElementPattern<PsiElement> =
-        StandardPatterns.or(
-            psiElement(TSTypes.IDENT)
-                .and(psiElement().afterSiblingSkipping(
+        psiElement(TSTypes.IDENT)
+            .and(
+                psiElement().afterLeafSkipping(
                     psiElement(TokenType.WHITE_SPACE),
-                    psiElement(TSTypes.DATABLOCK)))
-        )
+                    psiElement(TSTypes.DATABLOCK)
+                )
+            )
 
     private fun inSingleton(): ElementPattern<PsiElement> =
-        StandardPatterns.or(
-            psiElement(TSTypes.IDENT)
-                .and(psiElement().afterSiblingSkipping(
+        psiElement(TSTypes.IDENT)
+            .and(
+                psiElement().afterLeafSkipping(
                     psiElement(TokenType.WHITE_SPACE),
-                    psiElement(TSTypes.SINGLETON)))
-        )
+                    psiElement(TSTypes.SINGLETON)
+                )
+            )
 
     private fun inNew(): ElementPattern<PsiElement> =
-        StandardPatterns.or(
-            psiElement(TSTypes.IDENT)
-                .and(psiElement().afterSiblingSkipping(
+        psiElement(TSTypes.IDENT)
+            .and(
+                psiElement().afterLeafSkipping(
                     psiElement(TokenType.WHITE_SPACE),
-                    psiElement(TSTypes.NEW)))
-        )
+                    psiElement(TSTypes.NEW)
+                )
+            )
 }