+++ src/org/lukasj/idea/torquescript/engine/EngineApiService.kt
@@ -63,9 +63,9 @@ class EngineApiService(private val project: Project) {
 
     fun getFunctions() = cachedApi?.value?.getAllFunctions() ?: listOf()
     fun getStaticFunctions() = getFunctions().filter { it.isStatic }
-    fun findFunction(name: String) = getFunctions().firstOrNull { it.name == name }
+    fun findFunction(name: String) = getFunctions().firstOrNull { it.name.equals(name, ignoreCase = true) }
     fun getClasses() = cachedApi?.value?.getAllClasses() ?: listOf()
-    fun findClass(className: String) = getClasses().firstOrNull {it.name == className}
+    fun findClass(className: String) = getClasses().firstOrNull { it.name.equals(className, ignoreCase = true) }
     fun getMethods(className: String) = findClass(className)?.methods ?: listOf()
     fun findMethod(className: String, name: String) = getMethods(className)
         .firstOrNull { it.name == name }