+++ src/org/lukasj/idea/torquescript/psi/TSObjectDeclaration.kt
@@ -1,6 +1,10 @@
 package org.lukasj.idea.torquescript.psi
 
+import com.intellij.psi.util.PsiTreeUtil
+
 interface TSObjectDeclaration : TSNamedElement {
     fun getObjectName(): TSObjectName
-    fun getParentBlock(): TSParentBlock
+    fun getParentBlock(): TSParentBlock?
+    open fun getTypeName(): String =
+        PsiTreeUtil.findSiblingForward(this.firstChild, TSTypes.IDENT, null)!!.text
 }